refactor: mock API 구조 개선
- 실제 함수와 함수명, 파라미터 통일
- mock.ts에서 관심사별로 분리
